+++ Core/include/Acts/Utilities/IAxis.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Utilities/detail/AxisFwd.hpp"
+#include "Acts/Utilities/AxisFwd.hpp"
 
 #include <vector>
